subida final
La antigua trainningmanager ahora tiene un 2 añadido. La nueva es la web
ahora mismo más los cambios que tenía que hacer para el proyecto en sí

+++ trainningmanager/tarifas/cambiariva.php
@@ -1,13 +1,19 @@
 <?php
-if(isset($_POST['cambiariva'])) {
-	include("../php/funciones/function.php");
-	$ivaS=$_POST['iva'];
-	$iva=(double)$ivaS;
-	if($iva!=0){
-		ordensqlupdate("DELETE from iva;");
-        ordensqlupdate("INSERT into iva (iva) VALUES (".$iva.");");
+include("../php/funciones/function.php"); 
+comprueba();
+if($_POST['iva']>=0 && is_numeric($_POST['iva'])){
+	if(isset($_POST['cambiariva'])) {
+		include ('../Mobile-Detect/Mobile_Detect.php');
+		$ivaS=$_POST['iva'];
+		$iva=(double)$ivaS;
+		if($iva!=0){
+			ordensqlupdate("DELETE from iva;");
+	        ordensqlupdate("INSERT into iva (iva) VALUES (".$iva.");");
+		}
 	}
+	header("Location:tarifas.php?mensaje=okiva&iva=".$ivaS."");
+}else{
+	header("Location:tarifas.php?mensaje=erroriva");
 }
-header("Location:tarifas.php");
 ?>
